Aprendendo Sobre gerenciamento de Memoria heap

+++ src/Array/aula01/Aluno.java
@@ -1,8 +1,7 @@
 package Array.aula01;
 
 public class Aluno {
-
     String nome;
-    int idade;
+    double idade;
 
 }
